Remove unused properties
