+++ js/src/pages/markets/markets-dashboard/edit-market-modal/edit-primary-audience.scss
@@ -1,4 +1,4 @@
-.gla-edit-primary-feed {
+.gla-edit-primary-audience {
 
 	.woocommerce-tree-select-control__label {
 		color: $gray-900;
